license: Add referenced 2-Clause BSD License
This adds a `COPYING` file to the repository with the 2-Clause BSD
License that is referenced in `bip-witaddr.mediawiki`. It also adds a
link back to COPYING in `bip-witaddr.mediawiki` in the aforementioned
reference.

This also updates the created date from `2016-03-20` to `2017-03-20`. I
was thinking that maybe this was supposed to be referencing yesterday,
not yesterday, a year ago.